Move team to Leave Fragment
