+++ packages/frontend-2/components/project/page/settings/general/block/Leave/Leave.vue
@@ -1,6 +1,7 @@
 <template>
   <div>
     <ProjectPageSettingsBlock
+      v-if="canLeaveProject"
       background
       title="Leave Project"
       :icon="ArrowRightOnRectangleIcon"
@@ -22,7 +23,7 @@
       </div>
     </ProjectPageSettingsBlock>
     <ProjectPageSettingsGeneralBlockLeaveDialog
-      v-if="project"
+      v-if="canLeaveProject"
       v-model:open="showLeaveDialog"
       :project="project"
     />
@@ -33,18 +34,38 @@
 import { ArrowRightOnRectangleIcon } from '@heroicons/vue/24/outline'
 import { graphql } from '~~/lib/common/generated/gql'
 import type { ProjectPageSettingsGeneralBlockLeave_ProjectFragment } from '~~/lib/common/generated/gql/graphql'
+import { Roles } from '@speckle/shared'
 
 graphql(`
   fragment ProjectPageSettingsGeneralBlockLeave_Project on Project {
     id
     name
     role
+    team {
+      role
+      user {
+        ...LimitedUserAvatar
+        role
+      }
+    }
   }
 `)
 
-defineProps<{
-  project?: ProjectPageSettingsGeneralBlockLeave_ProjectFragment
+const props = defineProps<{
+  project: ProjectPageSettingsGeneralBlockLeave_ProjectFragment
 }>()
 
 const showLeaveDialog = ref(false)
+
+const { activeUser } = useActiveUser()
+
+const canLeaveProject = computed(() => {
+  if (!activeUser.value || !props.project.role) {
+    return false
+  }
+
+  const userId = activeUser.value.id
+  const owners = props.project.team.filter((t) => t.role === Roles.Stream.Owner)
+  return owners.length !== 1 || owners[0].user.id !== userId
+})
 </script>
